Swap ADJUSTSTACK{UP,DOWN] in AlphaInstrInfo.td
Alpha, just like pretty much every other architecture, has a stack that
grows down. This means we should adjust the stack down, not up, when we set up
a call frame. This is consistent with the code (but not the misleading
comment, which I removed) in AlphaFrameLowering.cpp.

I have no idea how it worked before that.