Don't mark R29 as reserved
The register coalescer assumes the reserved registers do not change during
the function execution, while in fact R29 is just an ordinary register
which may, for example, be clobbered by subroutine calls.

While we're at it, remove R29 from uses of BSR/JSR -- it's not used
by these instructions explicitly and I'm not sure why it was there
in the first place.