Mark NOP instructions as terminators
In snippets like this one, we should explicitly print the $BB0_5 block label:
    br $31,$BB0_5
    bis $31,$31,$31 ; <-- this is a NOP
$BB0_5:
    [...]

However, without the terminator flag for the NOP instruction,
AsmPrinter::isBlockOnlyReachableByFallthrough(),
which scans basic blocks from bottom to top, fails to detect the
unconditional branch and removes the label as a result.

Note: multiple consecutive NOPs at the bottom of a block are possible,
so technically only the last one is a terminator proper. But marking
them all is simpler, and shouldn't affect other passes, since NOP insertion
happens just before assembly is actually emitted.